Extract an isDayDisabled helper function
