+++ src/date_utils.js
@@ -5,3 +5,12 @@ export function isSameDay(moment1, moment2) {
     return !moment1 && !moment2;
   }
 }
+
+export function isDayDisabled(day, { minDate, maxDate, excludeDates, includeDates, filterDate } = {}) {
+  return (minDate && day.isBefore(minDate, "day")) ||
+    (maxDate && day.isAfter(maxDate, "day")) ||
+    (excludeDates && excludeDates.some(excludeDate => isSameDay(day, excludeDate))) ||
+    (includeDates && !includeDates.some(includeDate => isSameDay(day, includeDate))) ||
+    (filterDate && !filterDate(day.clone())) ||
+    false;
+}
