+++ src/day.jsx
@@ -1,7 +1,7 @@
 import moment from "moment";
 import React from "react";
 import classnames from "classnames";
-import { isSameDay } from "./date_utils";
+import { isSameDay, isDayDisabled } from "./date_utils";
 
 var Day = React.createClass({
   displayName: "Day",
@@ -31,13 +31,7 @@ var Day = React.createClass({
   },
 
   isDisabled() {
-    const { day, minDate, maxDate, excludeDates, includeDates, filterDate } = this.props;
-
-    return (minDate && day.isBefore(minDate, "day")) ||
-      (maxDate && day.isAfter(maxDate, "day")) ||
-      (excludeDates && excludeDates.some(excludeDate => this.isSameDay(excludeDate))) ||
-      (includeDates && !includeDates.some(includeDate => this.isSameDay(includeDate))) ||
-      (filterDate && !filterDate(day.clone()));
+    return isDayDisabled(this.props.day, this.props);
   },
 
   isInRange() {
